Add admin attribute to users
The user model was currently without a way to identify whether or
not someone is an administator for the website. Which is necessary
since only the admin of the website to be able to change information
on the site or post any blogs. So this attribute was added to users
via a migration for the users table to include a boolean for admin.