Add redis and bcrypt gems
When the app was initiated the bcrypt and redis gems were commented
out so they weren't actually setup during the original bundle install.
So both of the gems were uncommented and bundle install was rerun
so that these gems can be utilized by this application. THe BCrypt
gem is necessary for user authentication, especially since there
will be an adiminstator that will be posting information to the
website. BCrypt helps encrypt and secure passwords so that they
cannot be easily hacked and stolen. It also helps with the
authentication process.