Add multiple columns for password_digest and others
The users table did not have a place for their password to be saved
so the password_digest attribute was added, this must be specifically
named password_digest to work with the BCrypt gem. A remember digest
and reset digest were also added so that a password change for a user
can be implemented on the website if forgotten or if they want to.