Add validations for name, email, password, etc
The attributes for the user model were not being validated in any
way so a new user could be created without any information actually
being entered, which would create a user with blank attributes. We
do not want this so a presence validation was added for name, email,
password,and password_confirmation. There is also a uniqueness
validation for the email that a user enters so there aren't multiple
users with the same email.