Add default value to newsletter attribute
The newsletter attribute did not have a default value set for if
a new user happened to not select yes or no when asked if they
would like to receive the website's monthly newsletter. But I did
not want it to a required attribute that must be selected in order
for the user to be valid, so I defaulted the newsletter to true for
if someone happens to not select which option they want. May have to
change the default to false later on, but will keep it true for now.